DevTools release script updates:
1. Add README
2. Split release script into separate prepare, build, and publish scripts.
