+++ scripts/devtools/.gitignore
@@ -1,1 +1,2 @@
+.build-metadata
 .progress-estimator
